코드 정리 및 인코딩 변경

+++ Factorio-Imitation/AssemblingMachine1.h
@@ -24,7 +24,7 @@ public:
 		// 재료를 소모하는 함수
 		bool Consume(Recipe* recipe);
 
-		ItemInfo& operator[](int index);
+		ItemInfo& operator[](size_t index);
 	};
 
 private:
